Extract circle polyline calculation util
