+++ src/components/Map/StaticGoogleMap.js
@@ -5,45 +5,12 @@ import isEqual from 'lodash/isEqual';
 import polyline from '@mapbox/polyline';
 import { encodeLatLng, stringify } from '../../util/urlHelpers';
 import { lazyLoadWithDimensions } from '../../util/contextHelpers';
+import { circlePolyline } from '../../util/maps';
 import config from '../../config';
 
 const DEFAULT_COLOR = 'FF0000';
 const DEFAULT_STROKE_OPACITY = 0.3;
 const DEFAULT_FILL_OPACITY = 0.2;
-
-// Return polyline encoded list of points forming a circle
-// This algorithm is based on
-// https://stackoverflow.com/questions/7316963/drawing-a-circle-google-static-maps
-const getEncodedGMapCirclePoints = (lat, lng, rad, detail = 8) => {
-  const R = 6371;
-  const pi = Math.PI;
-
-  const _lat = lat * pi / 180;
-  const _lng = lng * pi / 180;
-  const d = rad / 1000 / R;
-
-  let points = [];
-  for (let i = 0; i <= 360; i += detail) {
-    const brng = i * pi / 180;
-
-    let pLat = Math.asin(
-      Math.sin(_lat) * Math.cos(d) + Math.cos(_lat) * Math.sin(d) * Math.cos(brng)
-    );
-    const pLng =
-      (_lng +
-        Math.atan2(
-          Math.sin(brng) * Math.sin(d) * Math.cos(_lat),
-          Math.cos(d) - Math.sin(_lat) * Math.sin(pLat)
-        )) *
-      180 /
-      pi;
-    pLat = pLat * 180 / pi;
-
-    points.push([pLat.toFixed(6), pLng.toFixed(6)]);
-  }
-
-  return polyline.encode(points);
-};
 
 // Extract color from string. Given value should be either with '#' (e.g. #FFFFFF') or without it.
 const formatColorFromString = color => {
@@ -80,7 +47,6 @@ const drawFuzzyCircle = (coordinatesConfig, center) => {
   }
 
   const { fillColor, fillOpacity, strokeColor, strokeWeight } = coordinatesConfig.circleOptions;
-  const { lat, lng } = center;
 
   const circleRadius = coordinatesConfig.coordinateOffset || 500;
   const circleStrokeWeight = strokeWeight || 1;
@@ -89,8 +55,8 @@ const drawFuzzyCircle = (coordinatesConfig, center) => {
   const circleFill = formatColorFromString(fillColor);
   const circleFillOpacity = convertOpacity(fillOpacity || DEFAULT_FILL_OPACITY);
 
-  //Encoded polyline string
-  const encodedPolyline = getEncodedGMapCirclePoints(lat, lng, circleRadius, 8);
+  // Encoded polyline string
+  const encodedPolyline = polyline.encode(circlePolyline(center, circleRadius));
 
   const polylineGraphicTokens = [
     `color:0x${circleStrokeColor}${circleStrokeOpacity}`,
